split download and processing of data into two different processes
