+++ .gitignore
@@ -25,3 +25,4 @@ dist-ssr
 *.sln
 *.sw?
 public/data/casos.json
+services/data/sheet.tsv
