+++ package.json
@@ -8,7 +8,7 @@
     "dev": "vite",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d dist",
-    "build": "node services/google-sheets.js > public/data/casos.json && tsc --noEmit && vite build",
+    "build": "node services/fetchAndSaveSheet.js > services/data/sheet.tsv && node services/google-sheets.cjs > public/data/casos.json && tsc --noEmit && vite build",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx --report-unused-disable-directives --max-warnings 5",
     "preview": "vite preview",
     "test": "vitest"
